к модулю D19

+++ board/views.py
@@ -1,12 +1,13 @@
-from django.http import HttpResponseRedirect, HttpResponse
-from django.views.generic import ListView, DetailView, CreateView, UpdateView, DeleteView, TemplateView
-from django.contrib.auth.mixins import LoginRequiredMixin, PermissionRequiredMixin
-from .models import Post, Category, Feedback, User
+from django.http import HttpResponseRedirect
+from django.views.generic import ListView, DetailView, CreateView, UpdateView, DeleteView
+from django.contrib.auth.mixins import LoginRequiredMixin
+from .models import Post, Feedback
 from board.forms import PostForm, FeedbackForm, UpdateProfileForm, UpdateUserForm
 from board.filters import FeedFilter
 from django.shortcuts import render, redirect
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
+from django.urls import reverse, reverse_lazy
 
 
 # Create your views here.
@@ -21,6 +22,7 @@ def profile(request):
             user_form.save()
             profile_form.save()
             messages.success(request, 'Ваш профиль отредактирован успешно')
+
             return redirect('profile')
     else:
         user_form = UpdateUserForm(instance=request.user)
@@ -43,23 +45,31 @@ class PostDetail(DetailView):
     pk_url_kwarg = 'id'
 
 
-# def get_user(request):
-#     user = request.user
-#     return user
 class PostCreate(LoginRequiredMixin, CreateView):
     form_class = PostForm
     model = Post
     template_name = 'post_create.html'
-
-    # redirect_field_name = '/posts/'
 
     def form_valid(self, form):
         user_post = form.save(commit=False)
         user_post.author = self.request.user
 
         redirect_to_post = super().form_valid(form)
-        # send_mail_new_post.apply_async([user_post.id], countdown=5)
         return redirect_to_post
+
+
+class PostUpdate(LoginRequiredMixin, UpdateView):
+    model = Post
+    form_class = PostForm
+    template_name = 'post_edit.html'
+    pk_url_kwarg = 'id'
+
+
+class PostDelete(LoginRequiredMixin, DeleteView):
+    model = Post
+    template_name = 'post_del.html'
+    pk_url_kwarg = 'id'
+    success_url = reverse_lazy('posts_list')
 
 
 class FeedCreate(LoginRequiredMixin, CreateView):
@@ -78,9 +88,12 @@ class FeedCreate(LoginRequiredMixin, CreateView):
         post_id = self.kwargs.get('id')
         user_reply.user = self.request.user
         user_reply.post = Post.objects.get(id=post_id)
-        user_reply.save()
-        # send_mail_new_post.apply_async([user_post.id], countdown=5)
-        return HttpResponseRedirect('/posts/')
+        print(user_reply.user, user_reply.post.author)
+        if user_reply.user != user_reply.post.author:
+            user_reply.save()
+            return HttpResponseRedirect('/posts/')
+        else:
+            return redirect(reverse('post_detail', args=[str(post_id)]))
 
 
 class FeedPost(LoginRequiredMixin, ListView):
@@ -90,46 +103,39 @@ class FeedPost(LoginRequiredMixin, ListView):
     paginate_by = 5
 
     def get_queryset(self):
-        queryset = super().get_queryset()
-
+        queryset = super().get_queryset().select_related('post').filter(post__author=self.request.user)
         self.filterset = FeedFilter(self.request.GET, queryset, request=self.request)
         return self.filterset.qs
 
-    def get_context_data(self, **kwargs):
+    def get_context_data(self, *args, **kwargs):
         context = super().get_context_data(**kwargs)
         context['filterset'] = self.filterset
         return context
 
 
+@login_required
 def feed_deleter(request, id):
     feedback = Feedback.objects.get(id=id)
-    print(feedback)
-    feedback.delete()
-    return redirect("feed_post")
+    author = feedback.post.author
+    user = request.user
+    if author == user:
+        feedback.delete()
+        messages.success(request, 'Сообщение удалено!')
+        return redirect("feed_post")
+    else:
+        messages.error(request, 'Вы не можете удалить данное сообщение!')
+        return redirect("feed_post")
 
 
+@login_required
 def feed_updater(request, id):
     feedback = Feedback.objects.get(id=id)
     if feedback.accept == False:
         print(feedback)
         feedback.accept = True
         feedback.save()
+        messages.success(request, 'Сообщение принято успешно!')
         return redirect("feed_post")
     else:
-        return HttpResponse('<h2>ВЫ УЖЕ ПРИНИМАЛИ ДАННОЕ СООБЩЕНИЕ</h2>')
-
-# class FeedPostList(LoginRequiredMixin, ListView):
-#     model = Feedback
-#     template_name = 'feed_post.html'
-#     context_object_name = 'feed'
-#     paginate_by = 5
-#
-#     def get_queryset(self):
-#         queryset = super().get_queryset()
-#         post_id = self.kwargs.get('id')
-#         return queryset.filter(post=post_id)
-#
-#     def get_context_data(self, **kwargs):
-#         context = super().get_context_data(**kwargs)
-#         context['post'] = Post.objects.get(id=kwargs.get('id'))
-#         return context
+        messages.error(request, 'Вы уже принимали данное сообщение!')
+        return redirect("feed_post")
